requirements.txt
numpy>=1.18.2
cython>=0.28.5
tensorboard>=1.8.0      
tensorboardX>=1.6
cython>=0.28.5
imageio>=2.3.0
scikit-image>=0.14.0
pydensecrf>=1.0rc2  
scipy==1.1.0
opencv-python>=3.4.2.17  
pandas>=0.23.4
Pillow>=5.2.0
torch>=0.4.1
torchvision>=0.2.1
